feat: Home page removed. Index is home page now
